% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{Sample Function}
\usage{
foo(arg1 = 0, ...)
}
\arguments{
\item{arg1}{Argument description.}

\item{...}{Additional arguments.}
}
\value{
The value of \code{arg1}.
}
\description{
Sample function.
}
\details{
Sample function.
}
\examples{
library(psych)

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
